fixed tests by moving admin.js into app, changing bin file to a wrapper for run
